OMC Final Code v1.5
> Parpadeo de los leds
> Modificación falla por corriente

+++ OMC_WIFI_CODE/OMC_WIFI_FInalCode/OMC_WIFI_Codev1.5/OMC_WIFI_GlobalVariables.cpp
@@ -56,7 +56,7 @@ int corrFail = 0;
 
 uint8_t voltSup;     // Máximo voltaje permitido
 uint8_t voltInf;     // Mínimo voltaje permitido
-uint8_t corrSup;       // Máxima corriente permitida
+float corrSup;       // Máxima corriente permitida
 
 uint8_t tiempoRecuperacion = 20;      // Tiempo requerido permitir paso de corriente luego de una falla o un reinicio (segundos)
